Refactor create command (#542)
